♻️ Complete redux refactor
