+++ package.json
@@ -22,7 +22,7 @@
     "start": "npm run develop",
     "serve": "gatsby serve",
     "lint": "eslint .",
-    "lint:fix": "eslint ./src --fix",
+    "lint:fix": "eslint --ignore-path .gitignore . --fix",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
   },
   "husky": {
