+++ src/components/dashboard.js
@@ -19,7 +19,7 @@ import ListItemText from '@material-ui/core/ListItemText'
 import InboxIcon from '@material-ui/icons/MoveToInbox'
 import MailIcon from '@material-ui/icons/Mail'
 import { connect } from 'react-redux'
-import { toggleSidebar } from 'state/actions'
+import { toggleSidebar } from 'actions'
 import PropTypes from 'prop-types'
 import Avatar from '@material-ui/core/Avatar'
 import { navigate } from '../../node_modules/gatsby-link/index'
