+++ src/pages/profile.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import UserLayout from '@/userLayout'
 import { connect } from 'react-redux'
-import { loadUserInfo, toggleEditProfileModal } from 'state/actions'
+import { loadUserInfo, toggleEditProfileModal } from 'actions'
 import Avatar from '@material-ui/core/Avatar'
 import Card from '@material-ui/core/Card'
 import CardActions from '@material-ui/core/CardActions'
