+++ src/pages/workflows.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import UserLayout from '@/userLayout'
 import { connect } from 'react-redux'
-import { loadUserWorkflows, toggleWorkflowModal } from 'state/actions'
+import { loadUserWorkflows, toggleWorkflowModal } from 'actions'
 import Button from '@material-ui/core/Button'
 import AddUserWorkflow from '@/addUserWorkflow'
 import UserWorkflows from '@/userWorksflows'
