+++ src/reducers/index.js
@@ -1,0 +1,13 @@
+import { combineReducers } from 'redux'
+
+import uiReducer from 'reducers/uiReducer'
+import workflowsReducer from 'reducers/workflowsReducer'
+import workflowReducer from 'reducers/workflowReducer'
+import userReducer from 'reducers/userReducer'
+
+export default combineReducers({
+  user: userReducer,
+  workflows: workflowsReducer,
+  workflow: workflowReducer,
+  ui: uiReducer,
+})
