--- src/state/store.js
@@ -1,9 +1,0 @@
-import { composeWithDevTools } from 'redux-devtools-extension'
-import { createStore, applyMiddleware } from 'redux'
-import thunk from 'redux-thunk'
-import rootReducer from './reducers'
-
-export default createStore(
-  rootReducer,
-  composeWithDevTools(applyMiddleware(thunk))
-)
